chore(secure-routes): use jwt authentication token to secure routes
- ensure user pass a token in header when making queries

[Delivers #157802731]
