+++ server/src/controllers/RequestController.js
@@ -71,7 +71,6 @@ const RequestController = class {
   static updateRequest(req, res) {
     const requestId = parseInt(req.params.requestId, 10);
     const userId = parseInt(req.query.userId, 10);
-    const { isAdmin } = req.query;
     if (!requestId || !userId) {
       return res.status(400).send({ message: 'missing required field' });
     }
@@ -82,36 +81,20 @@ const RequestController = class {
           id: requestId,
         },
       })
-      /* eslint-disable consistent-return */
       .then((request) => {
         // users should not be able to modify the status of a request
-        if (!isAdmin) {
-          return requests
-            .update(
-              request,
-              {
-                category: req.body.category || request.category,
-                description: req.body.description || request.description,
-                department: req.body.department || request.department,
-                urgency: req.body.urgency || request.urgency,
-              },
-            )
-            .then(newRequest => res.status(200).send(newRequest))
-            .catch(error => res.status(500).send(error));
-        }
-
-        // admin should be able to modify only the status of a request
-        if (isAdmin) {
-          return requests
-            .update(
-              request,
-              {
-                status: req.body.status || request.status,
-              },
-            )
-            .then(newRequest => res.status(200).send(newRequest))
-            .catch(error => res.status(500).send(error));
-        }
+        return requests
+          .update(
+            request,
+            {
+              category: req.body.category || request.category,
+              description: req.body.description || request.description,
+              department: req.body.department || request.department,
+              urgency: req.body.urgency || request.urgency,
+            },
+          )
+          .then(newRequest => res.status(200).send(newRequest))
+          .catch(error => res.status(500).send(error));
       })
       .catch(error => res.status(404).send(error));
   }
