+++ server/src/routes/index.js
@@ -1,4 +1,5 @@
 import multer from 'multer';
+import jwt from 'jsonwebtoken';
 import UsersController from './../controllers/UsersController';
 import RequestController from './../controllers/RequestController';
 import ContactController from './../controllers/ContactController';
@@ -23,19 +24,38 @@ const Routes = class {
       res.status(200).send({ message: 'welcome to the maintenance trackers' });
     });
 
-    // routes for users
-    app.post('/api/v1/users/signup', usersUpload.single('profile-photo'), this.UsersController.signup);
-    app.post('/api/v1/users/signin', this.UsersController.signin);
+    // secure api
+    app.use('/secure/', (req, res, next) => {
+      const { token } = req.headers;
+      /* eslint-disable no-unused-vars */
+      const promise = new Promise((resolve, reject) => {
+        if (token) {
+          jwt.verify(token, process.env.SECRET_KEY, (err, decode) => {
+            if (err) {
+              resolve(res.status(401).send('Invalid Token'));
+            } else {
+              resolve(next());
+            }
+          });
+        } else {
+          resolve(res.status(402).send('Please send a token'));
+        }
+      });
+      return promise;
+    });
+    // routes for us
+    app.post('/api/v1/auth/signup', usersUpload.single('profile-photo'), this.UsersController.signup);
+    app.post('/api/v1/auth/signin', this.UsersController.signin);
     // routes for requests model
-    app.post('/api/v1/users/requests', this.RequestController.addRequest);
-    app.get('/api/v1/users/requests/:requestId', this.RequestController.getOneRequest);
-    app.get('/api/v1/users/requests', this.RequestController.getAllRequests);
-    app.put('/api/v1/users/requests/:requestId', this.RequestController.updateRequest);
-    app.delete('/api/v1/users/requests/:requestId', this.RequestController.deleteRequest);
+    app.post('/secure/api/v1/users/requests', this.RequestController.addRequest);
+    app.get('/secure/api/v1/users/requests/:requestId', this.RequestController.getOneRequest);
+    app.get('/secure/api/v1/users/requests', this.RequestController.getAllRequests);
+    app.put('/secure/api/v1/users/requests/:requestId', this.RequestController.updateRequest);
+    app.delete('/secure/api/v1/users/requests/:requestId', this.RequestController.deleteRequest);
 
     // routes for contacts model
-    app.post('/api/v1/contacts', this.ContactController.addMessage);
-    app.get('/api/v1/contacts', this.ContactController.getMessages);
+    app.post('/secure/api/v1/contacts', this.ContactController.addMessage);
+    app.get('/secure/api/v1/contacts', this.ContactController.getMessages);
   }
 };
