+++ server/trans/test/serverTest.js
@@ -1,5 +1,9 @@
 'use strict';
 
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
 var _chai = require('chai');
 
 var _chai2 = _interopRequireDefault(_chai);
@@ -12,27 +16,19 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var server = new _app2.default();
 var expect = _chai2.default.expect;
+exports.default = describe('Server', function () {
+  describe('unit test', function () {
+    it('should export a function', function () {
+      expect(_app2.default).to.be.a('function');
+    });
 
-// enforce test to run only in test env
+    it('should to be an object', function () {
+      expect(server).to.be.a('Object');
+    });
 
-if (process.env.NODE_ENV !== 'test') {
-  /* eslint-disable no-console */
-  console.log('can\'t run test in non test env. you are in ' + process.env.NODE_ENV + ' environment');
-} else {
-  describe('Server', function () {
-    describe('unit test', function () {
-      it('should export a function', function () {
-        expect(_app2.default).to.be.a('function');
-      });
-
-      it('should to be an object', function () {
-        expect(server).to.be.a('Object');
-      });
-
-      it('server should be an instanceOf Server', function () {
-        expect(server).to.be.an.instanceOf(_app2.default);
-      });
+    it('server should be an instanceOf Server', function () {
+      expect(server).to.be.an.instanceOf(_app2.default);
     });
   });
-}
+});
 //# sourceMappingURL=serverTest.js.map
